requirements.txt
apscheduler>=3.11.0
email-validator>=2.2.0
flask-cors>=6.0.1
flask>=3.1.1
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
numpy
openai>=1.95.0
openpyxl>=3.1.5
pandas>=2.3.1
psycopg2-binary>=2.9.10
requests>=2.32.4
xlsxwriter>=3.2.5
xlrd>=2.0.2
python-dotenv